% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{extract_best_min_loss}
\alias{extract_best_min_loss}
\title{Extracting min loss run}
\usage{
extract_best_min_loss(RESULTS)
}
\arguments{
\item{RESULTS}{result list}
}
\description{
Extracting min loss run
}
