% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{multi_runs}
\alias{multi_runs}
\title{TRIFASE multi run}
\usage{
multi_runs(
  RUNS = 50,
  CORES = 1,
  X,
  distX = NULL,
  disttX = NULL,
  S,
  K,
  R,
  tau,
  phi,
  kernel = c("exponential", "none"),
  n_iter = 1000,
  seed = 213,
  nugget = 1e-08,
  warm = FALSE,
  swipes = 3,
  estimate_tau = TRUE,
  threshold = 0.001,
  upd_row = c("C", "S"),
  upd_col = c("S", "C", "A")
)
}
\description{
TRIFASE multi run
}
