% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{single_run}
\alias{single_run}
\title{Trifase}
\usage{
single_run(
  X,
  seed,
  L,
  Linv,
  Dinv,
  iKER,
  tau,
  K,
  R,
  upd_row,
  upd_col,
  n_iter,
  nugget,
  swipes,
  estimate_tau,
  warm = F,
  HCdX,
  HCdtX,
  threshold
)
}
\arguments{
\item{X}{...}

\item{seed}{...}

\item{L}{...}

\item{Linv}{...}

\item{Dinv}{...}

\item{iKER}{...}

\item{tau}{...}

\item{K}{...}

\item{R}{...}

\item{upd_row}{...}

\item{upd_col}{...}

\item{n_iter}{...}

\item{nugget}{...}

\item{swipes}{...}

\item{estimate_tau}{...}

\item{warm}{...}

\item{HCdX}{...}

\item{HCdtX}{...}

\item{threshold}{...}
}
\description{
Trifase
}
