\name{TRIFASE-package}
\alias{TRIFASE-package}
\alias{TRIFASE}
\docType{package}
\title{
\packageTitle{TRIFASE}
}
\description{
\packageDescription{TRIFASE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TRIFASE}
\packageIndices{TRIFASE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TRIFASE}

Maintainer: \packageMaintainer{TRIFASE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
